Blocking probability

+++ src/types.h
@@ -5,7 +5,7 @@
 
 using Time = double;
 using Duration = double;
-using Size = double;
+using Size = int64_t;
 using Intensity = double;
 using Uuid = uint64_t;
